Fixed issue : Problem with ReactRouter / Github Pages / HashRouter vs BrowserRouter component
Ok, setting up a Github Pages for this project is just for fun but it messed up a little bit cause of this "HashRouter" component.
So i prefer to switch back to new BrowserRouter on the main branch cause this is what is going to be shown at the exam.